moved project into separate folder
